Move db from common
