+++ src/main/scala/com/softwaremill/realworld/articles/comments/CommentsServerEndpoints.scala
@@ -2,8 +2,8 @@ package com.softwaremill.realworld.articles.comments
 
 import com.softwaremill.realworld.articles.comments.api.*
 import com.softwaremill.realworld.common.*
-import com.softwaremill.realworld.common.db.{Db, DbConfig}
-import com.softwaremill.realworld.http.ErrorMapper.defaultErrorsMappings
+import com.softwaremill.realworld.common.ErrorMapper.defaultErrorsMappings
+import com.softwaremill.realworld.db.{Db, DbConfig}
 import io.getquill.SnakeCase
 import sttp.model.StatusCode
 import sttp.tapir.generic.auto.*
