+++ src/main/scala/com/softwaremill/realworld/common/CustomDecodeFailureHandler.scala
@@ -1,4 +1,4 @@
-package com.softwaremill.realworld
+package com.softwaremill.realworld.common
 
 import com.softwaremill.realworld.common.ValidationFailed
 import sttp.model.{Header, StatusCode}
