+++ src/main/scala/com/softwaremill/realworld/users/UserRow.scala
@@ -1,5 +1,6 @@
-package com.softwaremill.realworld.common.db
+package com.softwaremill.realworld.users
 
+//TODO Rows only in repositories, probably make copy to Article
 case class UserRow(
     userId: Int,
     email: String,
